Changes required for Safari to be happy with modules
