+++ webroot/index.html
@@ -184,11 +184,13 @@
   <script type="module" src="js/message.js?v=2"></script>
   <script src="js/social.js"></script>
   <script src="js/components.js"></script>
-  <script src="js/player.js"></script>
   <script type="module">
       import Owncast from './js/app.js';
-      const app = new Owncast();
-      app.init();
+
+      (function () {
+        const app = new Owncast();
+        app.init();
+      })();
   </script>
 
   <noscript>
