+++ webroot/js/utils.js
@@ -1,62 +1,11 @@
 
 const LOCAL_TEST = window.location.href.indexOf('localhost:') >= 0;
-
-
 const URL_PREFIX = LOCAL_TEST ? 'http://localhost:8080' : ''; 
-const URL_STATUS = `${URL_PREFIX}/status`;
-const URL_CHAT_HISTORY = `${URL_PREFIX}/chat`;
-const URL_STREAM = `${URL_PREFIX}/hls/stream.m3u8`;
 
 const POSTER_DEFAULT = `${URL_PREFIX}/img/logo.png`;
 const POSTER_THUMB = `${URL_PREFIX}/thumbnail.jpg`;
 
-const URL_CONFIG = `${URL_PREFIX}/config`;
-
 const URL_OWNCAST = 'https://github.com/gabek/owncast'; // used in footer
-
-
-// Video setup
-const VIDEO_ID = 'video';
-const VIDEO_SRC = {
-  src: URL_STREAM,
-  type: 'application/x-mpegURL',
-};
-const VIDEO_OPTIONS = {
-  autoplay: false,
-  liveui: true, // try this
-  preload: 'auto',
-  html5: {
-    vhs: {
-      // used to select the lowest bitrate playlist initially. This helps to decrease playback start time. This setting is false by default.
-      enableLowInitialPlaylist: true,
-
-    }
-  },
-  liveTracker: {
-    trackingThreshold: 0,
-  },
-  sources: [VIDEO_SRC],
-};
-
-// local storage keys for chat
-const KEY_USERNAME = 'owncast_username';
-const KEY_AVATAR = 'owncast_avatar';
-const KEY_CHAT_DISPLAYED = 'owncast_chat';
-const KEY_CHAT_FIRST_MESSAGE_SENT = 'owncast_first_message_sent';
-
-const TIMER_STATUS_UPDATE = 5000; // ms
-const TIMER_DISABLE_CHAT_AFTER_OFFLINE = 5 * 60 * 1000; // 5 mins
-const TIMER_STREAM_DURATION_COUNTER = 1000;
-
-const TEMP_IMAGE = 'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7';
-
-const MESSAGE_OFFLINE = 'Stream is offline.';
-const MESSAGE_ONLINE = 'Stream is online';
-
-const CHAT_INITIAL_PLACEHOLDER_TEXT = 'Type here to chat, no account necessary.';
-const CHAT_PLACEHOLDER_TEXT = 'Message';
-const CHAT_PLACEHOLDER_OFFLINE = 'Chat is offline.';
-
 
 function getLocalStorage(key) {
   try {
